reorganize ENSKit API
- rename `resolve` to `contenthash`
- preserve `contenthash`, `text`, `avatar` as convenience methods in ENSKit
  - use these procedures when only a quick result is needed
- extract more specific procedures into `ENSResolver`
  - use this interface when finer control and error handling is required
