+++ Tests/ENSKitTests/ENSKitTests.swift
@@ -5,32 +5,33 @@ final class ENSKitTests: XCTestCase {
     let main = ENSKit()
 
     func testAvatarURL() async throws {
-        if let avatar = try await main.getAvatar(name: "vitalik.eth") {
-            let avatarURL = try await main.getAvatarImageURL(avatar: avatar)
+        if let resolver = try await main.resolver(name: "vitalik.eth"),
+           let avatar = try await resolver.getAvatar(),
+           let avatarURL = try await resolver.getAvatarImageURL(from: avatar) {
             XCTAssertEqual(avatarURL, URL(string: "ipfs://ipfs/QmSP4nq9fnN9dAiCj42ug9Wa79rqmQerZXZch82VqpiH7U/image.gif")!)
         } else {
             XCTFail()
         }
     }
 
-    func testResolveIPFS() async throws {
-        let vitalik = try await main.resolve(name: "vitalik.eth")
-        // last updated: 2022-05-27
-        XCTAssertEqual(vitalik, URL(string: "ipfs://QmbKu58pyq3WRgWNDv9Zat39QzB7jpzgZ2iSzaXjwas4MB"))
+    func testIPFSContenthash() async throws {
+        let vitalik = await main.contenthash(name: "vitalik.eth")
+        // last updated: 2022-08-15
+        XCTAssertEqual(vitalik, URL(string: "ipfs://QmQhCuJqSk9fF58wU58oiaJ1qbZwQ1eQ8mVzNWe7tgLNiD"))
     }
 
-    func testResolveIPNS() async throws {
-        let planetable = try await main.resolve(name: "planetable.eth")
+    func testIPNSContenthash() async throws {
+        let planetable = await main.contenthash(name: "planetable.eth")
         XCTAssertEqual(planetable, URL(string: "ipns://k51qzi5uqu5dgv8kzl1anc0m74n6t9ffdjnypdh846ct5wgpljc7rulynxa74a"))
     }
 
-    func testResolveIPNSWithDNSLink() async throws {
-        let uniswap = try await main.resolve(name: "uniswap.eth")
+    func testIPNSWithDNSLinkContenthash() async throws {
+        let uniswap = await main.contenthash(name: "uniswap.eth")
         XCTAssertEqual(uniswap, URL(string: "ipns://app.uniswap.org"))
     }
 
-    func testResolveText() async throws {
-        let coaEmail = try await main.text(name: "coa.eth", key: "email")
+    func testText() async throws {
+        let coaEmail = await main.text(name: "coa.eth", key: "email")
         XCTAssertEqual(coaEmail, "hello@carloscar.com")
     }
 }
